+++ landingzones/tests/test_views_taskflow.py
@@ -947,9 +947,9 @@ class TestZoneMoveView(
         self.assertEqual(len(self.zone_coll.data_objects), 0)
         self.assertEqual(len(self.assay_coll.data_objects), 2)
 
-    @override_settings(LANDINGZONES_ZONE_MOVE_VERIFY=True)
     def test_post_move_verify(self):
         """Test POST with verification enabled"""
+        app_settings.set(APP_NAME, 'zone_move_verify_site', True)
         irods_obj = self.make_irods_object(self.zone_coll, TEST_OBJ_NAME)
         self.make_checksum_object(irods_obj)
         self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)